requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
arg==0.0.1
args==0.1.0
asgiref==3.5.2
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
bleach==5.0.1
boto3==1.26.19
botocore==1.29.19
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
clint==0.5.1
colorama==0.4.6
commonmark==0.9.1
coverage==6.5.0
coveralls==3.3.1
cryptography==38.0.4
curdling==0.4.1
defusedxml==0.7.1
distlib==0.1.2
dj-database-url==1.0.0
Django==4.1.3
django-appconf==1.0.5
django-countries==7.4.2
django-environ==0.9.0
django-excel==0.0.10
django-extensions==3.2.1
django-filter==22.1
django-heroku==0.3.1
django-imagekit==4.1.0
django-ipware==4.0.2
django-mathfilters==1.0.0
django-qr-code==3.1.1
django-rest-framework==0.1.0
django-rest-knox==4.2.0
django-storages==1.13.1
djangorestframework==3.14.0
docopt==0.6.2
docutils==0.19
frozenlist==1.3.3
funcsigs==1.0.2
geoip2==4.6.0
gunicorn==20.1.0
heroku==0.1.1
html2image==2.0.1
html2pdf==0.1
htmlwebshot==0.1.2
httpretty==1.1.4
idna==3.4
imgkit==1.2.2
importlib-metadata==5.1.0
jaraco.classes==3.2.3
jmespath==1.0.1
keyring==23.11.0
lml==0.1.0
m2r==0.3.1
maxminddb==2.2.0
mistune==0.8.4
mock==4.0.3
more-itertools==9.0.0
multidict==6.0.2
mysqlclient==2.1.1
oauthlib==3.2.2
paystackapi==2.1.0
pbr==5.11.0
pilkit==2.0
Pillow==9.3.0
pkginfo==1.9.2
psycopg2==2.9.5
pycparser==2.21
pydantic==1.10.2
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
Pygments==2.13.0
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.6
pywin32-ctypes==0.2.0
readme-renderer==37.3
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==12.6.0
s3transfer==0.6.0
segno==1.5.2
simplejson==3.18.0
six==1.16.0
sqlparse==0.4.3
sure==2.0.0
texttable==1.6.7
tqdm==4.64.1
twine==4.0.1
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.13
webencodings==0.5.1
whitenoise==6.2.0
yarl==1.8.1
zipp==3.11.0
